Add basic game classes
We'll have a game keep track of all of the pieces in play. Ugh… I gotta implement physics and whatnot, though. That's tricky. The plan is to pass the board to something whenever we need to do an action. So when a character attacks, for instance, we'll know the entire context, and we can directly interact with the objects around the character.